Move qjit device
